requirements.txt
Django
djangorestframework
django-cors-headers
python-decouple
psycopg2-binary
cryptography
drf-yasg
celery
redis